-enh cleaned up the I2C base
