+++ modules/hal_i2c/src/I2CHelper.hpp
@@ -26,10 +26,18 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  ***********************************************************************************************/
 
-#ifndef I2CADDR_HPP
-#define I2CADDR_HPP
+#ifndef I2CHELPER_HPP
+#define I2CHELPER_HPP
 
 #include <stdint.h>
+
+#ifdef __arm__
+#include <Wire.h>
+#define I2C_PERIPHERAL_T TwoWire *
+#else
+#include "i2c_peripheral_mock.hpp"
+#define I2C_PERIPHERAL_T I2CPeripheralMock *
+#endif
 
 namespace hal::i2c
 {
@@ -40,6 +48,26 @@ namespace hal::i2c
         BREATHING_MODULE = 1,
         SENSOR_HUB = 2
     };
+    /// @brief error codes for i2c, should be extention of existing error handler enum
+    enum ErrorCode
+    {
+        SUCCESS,
+        ZERO_BYTES_WRITTEN,
+        NOT_ENOUGH_BUFFER_SPACE,
+        EOT_DATA_TOO_LONG,
+        EOT_NACK_ADDRESS,
+        EOT_NACK_DATA,
+        EOT_OTHER_ERROR,
+        EOT_ERROR_UNKNOWN
+
+    };
+
+    /// @brief Possible speed to use for the I2C protocol
+    typedef enum
+    {
+        kI2cSpeed_100KHz = 100000,
+        kI2cSpeed_400KHz = 400000,
+    } I2CSpeed;
 }
 
 #endif
