+++ modules/hal_i2c/src/I2CMasterBase.cpp
@@ -30,14 +30,61 @@
 
 namespace hal::i2c
 {
-    void I2CMasterBase::ReadFrom(I2CAddr i2c_addr, uint8_t *buffer, uint8_t num_of_bytes)
+    ErrorCode I2CMasterBase::ReadBytes(I2CAddr i2c_addr, uint8_t *buffer, uint8_t num_of_bytes)
     {
-        ReadBytes(buffer, num_of_bytes, i2c_addr);
+        uint8_t error;
+        error = i2c_peripheral_->requestFrom(i2c_addr, num_of_bytes, true);
+        if (num_of_bytes < error)
+        {
+            // The amount of bytes is more than the buffer can hold
+            // TODO: handle error (make buffer bigger?)
+            return NOT_ENOUGH_BUFFER_SPACE;
+        }
+
+        error = i2c_peripheral_->readBytes(buffer, num_of_bytes);
+        if (!error)
+        {
+            // No bytes where written to the buffer
+            // TODO: handle error
+            return ZERO_BYTES_WRITTEN;
+        }
+
+        return SUCCESS;
     }
 
-    void I2CMasterBase::SendTo(I2CAddr i2c_addr, uint8_t *buffer, uint8_t num_of_bytes)
+    ErrorCode I2CMasterBase::SendBytes(I2CAddr i2c_addr, uint8_t *buffer, uint8_t num_of_bytes)
     {
-        SendBytes(buffer, num_of_bytes, i2c_addr);
+        uint8_t error;
+        i2c_peripheral_->beginTransmission(i2c_addr);
+        i2c_peripheral_->write(buffer, num_of_bytes);
+        error = i2c_peripheral_->endTransmission(true);
+        ErrorCode errorCode = HandleEndTransmisstionError(error);
+        if (errorCode != SUCCESS)
+        {
+            // TODO: handle error(?)
+            return errorCode;
+        }
+
+        return errorCode;
+    }
+
+    ErrorCode I2CMasterBase::HandleEndTransmisstionError(uint8_t code)
+    {
+        switch (code)
+        {
+        case 0:
+            return SUCCESS;
+        case 1:
+            return EOT_DATA_TOO_LONG;
+        case 2:
+            return EOT_NACK_ADDRESS;
+        case 3:
+            return EOT_NACK_DATA;
+        case 4:
+            return EOT_OTHER_ERROR;
+        default:
+            return EOT_ERROR_UNKNOWN;
+        }
     }
 
 }
