+++ modules/hal_i2c/src/I2CSlaveBase.cpp
@@ -8,4 +8,9 @@ namespace hal::i2c
         i2c_peripheral_->onReceive(receiveEvent);
         i2c_peripheral_->onRequest(requestEvent);
     }
+
+    void I2CSlaveBase::ChangeAddress(I2CAddr new_i2c_address)
+    {
+        slave_addr_ = new_i2c_address;
+    }
 }
